# feat: add `unstable_renameRequire` to allow disabling the require rename transform