--- a/packages/metro-config/src/defaults/index.js
+++ b/packages/metro-config/src/defaults/index.js
@@ -135,7 +135,7 @@ const getDefaultValues = (projectRoot: ?string): ConfigT => ({
     unstable_dependencyMapReservedName: null,
     unstable_disableModuleWrapping: false,
     unstable_disableNormalizePseudoGlobals: false,
-    unstable_disableRequireRename: false,
+    unstable_renameRequire: true,
     unstable_compactOutput: false,
     unstable_workerThreads: false,
   },