--- a/packages/metro-transform-worker/src/index.js
+++ b/packages/metro-transform-worker/src/index.js
@@ -97,6 +97,8 @@ export type JsTransformerConfig = $ReadOnly<{
   unstable_compactOutput: boolean,
   /** Enable `require.context` statements which can be used to import multiple files in a directory. */
   unstable_allowRequireContext: boolean,
+  /** Disable renaming scoped `require` functions to `_$$_REQUIRE` which is an extraneous operation when serializing to iife (default). */
+  unstable_disableRequireRename: boolean,
 }>;
 
 export type {CustomTransformOptions} from 'metro-babel-transformer';
@@ -388,6 +390,7 @@ async function transformJS(
         importAll,
         dependencyMapName,
         config.globalPrefix,
+        config.unstable_disableRequireRename,
       ));
     }
   }