… packages/metro-config/src/__tests__/__snapshots__/loadConfig-test.js.snap