<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory. -->

## Summary

Renaming `require` to `_$$_REQUIRE` is an extraneous babel traversal that apparently is used internally at Meta for additional dependency extraction. Since the modules are scoped as iife functions, and there's no special convention to preserve native `require` syntax, this transform does not appear to add any value to standard projects. Per advice from @robhogan, I've added a flag to disable this transform, which we can enable by default in Expo CLI. 

Saves ~165ms when transforming the production react renderer module.

## Test plan