@EvanBacon 2024-03-06
Bundling our kitchen sink app (3301 modules):

`bun expo export -p ios --no-bytecode --max-workers=0 --no-minify -c `

- Before: 61,953ms
- After: 58,145ms

I'm finding I can only accurately measure savings with threading disabled. With threads enabled, the results can sometimes be slower, but are ~100ms faster (total average: 27,491ms).